chatbot/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
google-generativeai==0.6.0
gunicorn==20.1.0